Fixed username and password visibility

+++ theme/member-register.php
@@ -318,18 +318,16 @@
 	echo "  <div id=\"member-register\" class=\"rbform\">";
 	echo "	<p class=\"rbform-description\">".__("To Join Our Team please complete the application below.", rb_agency_interact_TEXTDOMAIN)."</p>";
 	echo "  <form method=\"post\" action=\"". $rb_agency_interact_WPURL ."/profile-register/talent\">\n";    				
-	if($rb_agencyinteract_option_useraccountcreation == 0){
+	if ($rb_agencyinteract_option_registerconfirm == 1 && $rb_agencyinteract_option_useraccountcreation == 1) {
 		echo "       <div id=\"profile-username\" class=\"rbfield rbtext rbsingle\">\n";
 		echo "       	<label for=\"profile_user_name\">". __("Username (required)", rb_agency_interact_TEXTDOMAIN) ."</label>\n";
 		echo "       	<div><input class=\"text-input\" name=\"profile_user_name\" type=\"text\" id=\"profile_user_name\" value=\""; if ( $error ) echo esc_html( $_POST['profile_user_name'], 1 ); echo "\" /></div>\n";
 		echo "       </div><!-- #rofile-username -->\n";
-				
-		if ($rb_agencyinteract_option_registerconfirm == 1) {
+
 		echo "       <div id=\"profile-password\" class=\"rbfield rbpassword rbsingle\">\n";
 		echo "       	<label for=\"profile_password\">". __("Password (required)", rb_agency_interact_TEXTDOMAIN) ."</label>\n";
 		echo "       	<div><input class=\"text-input\" name=\"profile_password\" type=\"password\" id=\"profile_password\" value=\""; if ( $error ) echo esc_html( $_POST['profile_password'], 1 ); echo "\" /></div>\n";
 		echo "       </div><!-- #profile-password -->\n";
-		}
 	}
 	echo "       <div id=\"profile-first-name\" class=\"rbfield rbtext rbsingle\">\n";
 	echo "       	<label for=\"profile_first_name\">". __("First Name", rb_agency_interact_TEXTDOMAIN) ."</label>\n";
